requirements.txt
absl-py==0.15.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.11.0
decorator==5.1.0
flatbuffers==1.12
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
ipython==7.29.0
jedi==0.18.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
Pyment==0.3.3
pyparsing==3.0.5
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
